Extract invites from main channel
