+++ services/app/apps/codebattle/lib/codebattle/invites_killer_server.ex
@@ -52,7 +52,7 @@ defmodule Codebattle.InvitesKillerServer do
         }
 
         CodebattleWeb.Endpoint.broadcast!(
-          "main",
+          "invites",
           "invites:expired",
           %{invite: data}
         )
